Add README

+++ FileOperations.h
@@ -10,7 +10,7 @@
 typedef struct file_part_s {
     size_t bit_amount;
     unsigned char *buffer;
-    int parity_size;
+    size_t parity_size;
     unsigned char *parity_file;
 } file_part;
